Use custom add_package to handle dependency checking in one line
This function looks similar to find_package, but it also specifies arguments to fallback to PkgConfig, and then external. This is essentially the bulk of all dependency handling, and a weird part of me wonders why it's not actually bundled into CMake (well, the 'external' stuff is definitely done on my own terms)

The only thing this doesn't really do is handle the output dependencies of find_package. I'd like to try to alias them all into the 'pkg_target' argument, but we'll see.