Support separating players in different sectors in remote play
The players of different remote users can now be separated in different sectors.

All sectors in a `GameSession` level that have at least 1 player in them are now updated.

All sectors of a `GameSession` level are now exposed to Squirrel at the same time, which means that each one now has its objects exposed under a sub-table under the `sector` table, which has the sector's name. This eliminates conflicts when having multiple active sectors, as well as allows for cross-sector scripting.

Additionally:
* `GameSession` now supports having multiple `SpawnRequest`s.
* `Sector::activate` was renamed to `Sector::spawn_players`, since it no longer does anything other than that. Init script is ran through a call from `GameSession` (anytime a player is spawned in a sector which had no players before), all sectors and their objects are exposed in `GameSession::setup()` and unexposed in `GameSession::leave()`. The `default.nut` script for the world is ran directly after exposing the sector.
* `Sector::s_current`, `Sector::get()` and `Sector::current()` were removed, since they do not allow for supporting multiple active sectors. Objects now refer to their parent sector through a pointer stored in `GameObject`.
* `Sector`s are now only exposed to Squirrel when in a `GameSession`, meaning that editor levels will no longer be exposed (not that there's ar eason they should be).