Fix nullptr crash in player status hud due to wrong player item pocket access
Fixes a crash if multiple players are in the game and the first player
does not have any bonus while other players have a bonus. Due to a wrong
access of m_bonus_sprites, a nullptr is returned for the sprite pointer
which causes a segfault when the hud tries to draw the sprite.