Fix relative paths with the Particle Editor
Relative paths make the game panic about insecure filenames when they contain '..'; when a custom particle file located outside the particles/ root folder was selected in the editor, and the object was used to 'Open in Particle Editor', the game would crash. The best fix would've been to dodge the need for relative paths entirely, but that would prevent the settings from opening by default in the particles/ folder. This is, I believe, the second best fix.