FreeBSD (and others?) error handling
FreeBSD and NetBSD require that execinfo needs to be linked. I didn't see anything regarding this in OpenBSD's (and dragonfly's??) but we can wait for someone to complain or an OpenBSD user sends/creates a patch. I was reading the manpages and all I tested on was FreeBSD. It works here for now.