Support X/Y position periods for autotile "alt-id"s (#3158)
Autotile "alt-id"s now support the "period-x" and "period-y" properties, which contain 2 unsigned integer values: first is the divider and second is the desired remainder after dividing the X/Y position of the tile, relative to its tilemap. If the desired remainder is matched, the condition is passed.

NOTE: For an "alt-id" to be chosen, all of its conditions ("period", random "weight") must be met.

This PR takes advantage of these new position periods to achieve proper castle autotiling with line tiles.