Editor comment (#3181)
The editor comment allows level designers to leave a note for themselves or other level designers collaborating on the same level. Similar to a code comment, it can be used to write to-do lists, fixmes, or notes.

This feature supports info box formatting, like what you'd see in the info block or the text scroller.