Fix addon MD5 buffer overflow (#3196)
The original code was using sizeof(so) which, on my machine, is returning 8. This causes the game to try to write to a point past so[32], causing a buffer overflow.
This PR simply hardcodes the size to 3 (as each char is always 1 byte large, theres 2 chars inserted via the format + 1 NULL char). This makes it not buffer overflow without breaking existing addon functionality.