Fixes #3079 Buttjump bonusblock runs script twice & unports trampoline
When someone buttjumps onto a bonus_block it calls the try_drop function
in which some objects on that function would call try_open, another
function on which it would run a script, then come back to try_drop and
run the script again since the condition to run the script didn't
include countdown, the bool that checks if try_open is called, so I
just added it to the condition.
Then for the trampoline error the add function had the position that
the trampoline is created and an enum (1 or 0, or respectivelly true
or false) which was incorrect and creating the wrong trampoline, so I
just changed from true, which created a stationary trampoline,
to a false, a portable one.

Closes #3079