Fixes #3208: bounce mriceblock to bonusblock by bumper doesn't go back
When a MrIceBlock is kicked into a trampoline, it bounces back facing
backwards. The collision_solid() function, when the MrIceBlock's state
is KICKED (which is the case), didn't expect to receive a hit from
the right while facing left, and vice-versa. I added that condition
and made it so it plays the collision sound and goes back, just like
normal, but skips the flipping orientation, as it should slide forward
and not backwards.

Closes #3208