Fixes #3209: Tux unable to stand on frozen badguy that is bouncing
Tux wasn't able to stand on or jump off a frozen badguy that was
bouncing of a trampoline, getting stuck in the falling animation
while on the frozen badguy. To solve this I added a verification
on Player::collision for when the badguy is frozen and the player
hits bottom. Inside the if I adapted the code on collision_solid
so that it performs similar to when the player collides with a
solid and the hit is on the bottom.

Closes #3209