Fixes #3210 - Stalactite landing on a trampoline (#3221)
The stalactite's death animation did not bounce on the trampoline
like other enemies do. Instead, it caused the trampoline to repeat
its bouncing sound and animation for a short period of time.
Now the stalactite's death animation bounces on the trampoline like
other enemies do, instead of remaining still on the trampoline.
To do this, I deleted a line in stalactite.cpp that changed the
stalactite's vertical speed to 0.

Closes #3210