Fixes #517 - Add a button with the ability to hide/single layers
The Supertux level editor did not have a button to hide
individual layers.
Now this button is at the end of each layer menu.
The button does not appear on the selected layer,
as it does not make sense to hide the selected layer.
When someone decides to hide a layer other than the
selected one and then switches to it, that layer is shown
again. In addition, after hiding a layer, if the hidden layer
menu is opened again, instead of the hide layer button,
a show layer button appears.
The default opacity of 0.5 was kept as the default,
so that users do not find it strange or have to press
this implemented button.

Closes #517

Co-authored-by: Filipe Oleacu <filipeoleacu04@tecnico.ulisboa.pt>