Fixes #3199 - Coins carry support for Owl enemy
Added the possibility for the owl to carry three new types
of objects, coin, coin rain and coin explosion.
Since the owl cannot visually carry a coin rain or a
coin explosion, then visually the owl for these types carries
a single coin, but then when the owl ungrabs the coin,
the behavior is of a coin rain or a coin explosion respectively.
The coin drop is activated when the tux approaches.

Closes #3199

Co-authored-by: Filipe Oleacu <filipeoleacu04@tecnico.ulisboa.pt>