Save temp level as stringstream for testing
Ok so, kind of a big hack. It's not really feasible at the moment to create a deep copy of the editor's level, since that involves either doing some fancy factory magic, or creating a million clone() virtual methods.

So for now, we save the level as a stringstream (viable std::ostream), then pass that level for the GameSession so it will end up loading the level from the stream. It's at least better than creating a temporary file.