Linked sprite support (#2753, #3418)
Custom light sprites can now be added to all objects, which inherit MovingSprite. Setting a custom color for the light sprite is also possible, as long as the object is not designed to use foreign color values (for example, from an option).

To add a light sprite with an RGB color of 0.5, 0.2, 0.25, in a .sprite file:

(linked-sprites
  (light "path/to/light/sprite" 0.5 0.2 0.25)
)

Keep in mind that specifying color values is optional. The entry can also end after specifying the file.

Additionally:

    Objects which utilize more than 1 sprite are now able to get the files of their additional sprites from the main one's linked-sprites list (the same block where light sprites can also be added, as shown above). This allows for expanded sprite customization for objects.

To add linked sprites, in a .sprite file:

(linked-sprites
  (key "path/to/sprite")
  (key2 "another/sprite/file")
)

Linked sprites can be added under sprite actions, too. If a sprite with the same key is found in the general linked sprites block and the action-specific one, the sprite specified by the action will be prioritized, thus overriding the default linked sprite file.

Sprites, which the specific object does not support, will be ignored.

    Sprite actions now support the flip-offset property, which determines how much the sprite should be vertically offset when flipped vertically.

-----

* Linked sprite support

Objects which utilize more than 1 sprite now get the files of their additional sprites from the main one's "linked-sprites" list.

This allows for expanded sprite customization for objects.

* `MovingSprite` light sprite support

Custom lightsprites can now be added to all objects, which inherit `MovingSprite`.

Setting a custom color for the lightsprite is also possible, as long as the object is not designed to use foreign color values (for example, from an option).

* `MovingSprite`: Use a map for linked sprites list, instead of struct

* "light" -> "on-light"

[ci skip]

* Undo unneeded changes to "red-potion.sprite"

[ci skip]

* `SpriteManager`: Dummy sprite data handling improvements

* Sprite action-specific linked sprites

Linked sprites can now be added to sprite actions too. If a sprite with the same key is found in the general linked sprites block and the action-specific one, the sprite specified by the action will be prioritized, thus overriding the default linked sprite file.

* Default actions for linked sprites

* Fix torch.sprite [ci skip]

* light is stupid i hate it

* dart light ooooo BUT ITS BLURRY WHAHAHA

* Un-hardcode linked sprite actions, allow for setting custom loop count

* Undo DiveMine hitbox changes [ci skip]

* Undo DiveMine hitbox changes [ci skip]

* Fix `Candle` custom light sprites not being configured

* Do not use `Dart` light sprite as `BadGuy::lightsprite`

* Fix `SpriteData` code duplication, remove sprite hardcoding for many `SpriteParticle`s

* `Crusher` now draws its custom light sprite

* Fix "Missing sprite" issues with some objects

* Replace `m_sprite->clone()` with a linked sprite in objects

Crushers and bricks now use linked sprites, instead of cloning their own sprite and setting actions on its clone.

* "light" -> "on-light", remove unneeded `set_blend`

* Remove more unneeded `set_blend(Blend::ADD)`

[ci skip]

* `WeakBlock`: `lightsprite` -> `m_burn_sprite`

[ci skip]

* Fix missing goldbomb "ticking-glow", attempt to fix dark `WeakBlock` "burn-light"

* Allow setting many custom sprites (light or not), use S-Expression objects

---------

Co-authored-by: Vankata453 <78196474+Vankata453@users.noreply.github.com>
Co-authored-by: MatusGuy <MatusGuy@users.noreply.github.com>

closes #2753