Fix wacky camera zoom (but only center zoom for now)
The logic of the old camera zoom was very out of wack. It made it impossible to
scroll into certain areas. This is mainly because the camera function oddly
expects you to call `update` when you zoom.

Zoom to pointer isn't implemented yet, but IMO this behavior is much preferable
to how it was before (which i'd argue was pretty much broken)

This also adds Camera::Mode::FREE which is basically Camera::Mode::NORMAL but
all methods ignore the camera boundaries (but, you can still update the camera
boundaries yourself, this is what the editor does as a bit of a lazy hack)

Signed-off-by: Swagtoy <me@ow.swag.toys>