Drop Ubuntu Touch support
Ubuntu Touch is pretty dead, even given that there is UBPorts (i think?).

Nobody is really motivated or around to maintain it anymore. It often causes
issues with newer SDL2 features and is a complete blocker for SDL3
support. Society is moving on to pmOS or whatever and such anyway; I am hopeful
that Flatpaks will ease that world up.

kubax and Semphris have acknowledged this decision. People are free to maintain
it in their own fork or as a patchset, but as it stands, nobody is interested,
so let's drop it.

Signed-off-by: Swagtoy <me@ow.swag.toys>