Count total Tux Dolls collected from levels in worldmap, require certain amount to unlock a world
After a level from a worldmap is finished, any collected Tux Dolls are added to the total count in the player status in the savegame of the world (if they are more than collected previously).

A contrib world can now have a required amount of Tux Dolls to be collected from Story Mode for it to be accessible.