Display music metadata in the level editor (#2674)
Fixes #2617.

When navigating or hovering over menu items in the level editor music object settings file browser, information about a song's author, license, and title will now be displayed in the menu item help text at the bottom of the screen.

This pull request adds:

- A software implementation
- author, license and title data for most SuperTux music files in data/music
- Updated authorship data in the game credits and data/AUTHORS
- A realignment of the display of menu item help text that fixes cropping of long (>4 line) menu item help texts that occur for some songs. (this is in a separate commit)

The author, license and title fields are option for music files. SuperTux simply omits the display of any fields that are not present in music files, and if none of these fields are defined for a music file (like christmas_theme.music, currently), then help text does not display for that song.

---------

Co-authored-by: provigz (Vankata453) <78196474+provigz@users.noreply.github.com>
Co-authored-by: Tobias Markus <tobbi.bugs@googlemail.com>