Count total Tux Dolls collected from levels in worldmap, require certain amount to unlock a world (#3451)
After a level from a worldmap is finished, any collected Tux Dolls are added to the total count in the player status in the savegame of the world (if they are more than collected previously).

A contrib world can now have a required amount of Tux Dolls to be collected from Story Mode for it to be accessible.

<img width="791" height="648" alt="Screenshot of locked world, requiring Tux Dolls" src="https://github.com/user-attachments/assets/8490eb08-701d-47b9-a3aa-bccf07eca364" />

The worldmap HUD now shows the amount of Tux Dolls collected, instead of coins.

A statistic for collected Tux Dolls in a level is now also shown in the statistics panel on worldmap, if at least 1 Tux Doll was collected from the level.

TODO: For a level designer to set a Tux Doll requirement for BI1, possibly also do modifications to info text in WTA.