worldmap: Add property to block direction (fork in the road)
In fork in the road scenarios, you have to rely on hacks to fade the path out so
Tux won't cross it. Now we reuse the special tile in a manner that you can just
block the direction. Then, if you need to unblock a path, you can simply set a
bitmask in Squirrel.

You put this tile over the level that you want a fork in the road for, and then
on clear, you set the worldmap state/bitmask (mostly what was done before, but
just a boolean) to allow Tux to go that direction.

Signed-off-by: Swagtoy <me@ow.swag.toys>