video: Remove windows pixel hack for now
I recall getting an issue report from it when the game was previously maximized?
I didn't get a lot of details and no followup, but I think it was related to
this since the video was shifted down, so let's not do this for now.

It was a fullscreen hack anyway, which you can _sorta_ fix by restarting the
game on windows i think, dunno.

Signed-off-by: Swagtoy <me@ow.swag.toys>