Hide native mouse cursor; move away from variable cursor draw
In the past, native mouse cursors (as well as custom native mouse cursors) never
actually hid themselves properly. This was because the previous logic was to
variably choose when to draw the mouse cursor, which was hardcoded into the
editor and such.

Instead, we're going to move that into ScreenManager (where it should've been)
and add a new m_visible property, so you instead choose when to draw depending
on if you're in a GameSession, worldmap, console, etc.

Signed-off-by: Swagtoy <me@ow.swag.toys>